refactor(frontend): atoms interface move to elements folder
