--- packages/frontend/src/components/atoms/interfaces.ts
@@ -1,11 +1,0 @@
-import { SetStateAction } from 'react';
-
-export interface InputParams {
-	type: string;
-	setValue: any;
-	placeHolder: string;
-	value: string;
-}
-export interface InputProps {
-	({ placeHolder, setValue, type, value }: InputParams): Element;
-}
