+++ packages/frontend/src/components/elements/interface.tsx
@@ -1,0 +1,9 @@
+export interface InputParams {
+  type: string;
+  setValue: any;
+  placeHolder: string;
+  value: string;
+}
+export interface InputProps {
+  ({ placeHolder, setValue, type, value }: InputParams): Element;
+}
